+++ src/main/java/com/erigitic/commands/SetBalanceCommand.java
@@ -11,13 +11,14 @@ import org.spongepowered.api.command.spec.CommandExecutor;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.event.cause.Cause;
 import org.spongepowered.api.event.cause.NamedCause;
+import org.spongepowered.api.service.economy.Currency;
 import org.spongepowered.api.text.Text;
 import org.spongepowered.api.text.format.TextColors;
 
 import java.math.BigDecimal;
 
 /**
- * Created by Erigitic on 8/7/2015.
+ * Created by Eric on 8/7/2015.
  */
 public class SetBalanceCommand implements CommandExecutor {
     private AccountManager accountManager;
@@ -31,13 +32,13 @@ public class SetBalanceCommand implements CommandExecutor {
         Player sender = ((Player) src).getPlayer().get();
         Player recipient = (Player) args.getOne("player").get();
         BigDecimal amount = new BigDecimal((String) args.getOne("amount").get()).setScale(2, BigDecimal.ROUND_DOWN);
-        Text symbol = accountManager.getDefaultCurrency().getSymbol();
+        Currency defaultCurrency = accountManager.getDefaultCurrency();
 
         TEAccount recipientAccount = (TEAccount) accountManager.getOrCreateAccount(recipient.getUniqueId()).get();
 
         recipientAccount.setBalance(accountManager.getDefaultCurrency(), amount, Cause.of(NamedCause.of("TotalEconomy", this)));
 
-        sender.sendMessage(Text.of(TextColors.GRAY, "You set ", recipient.getName(), "\'s balance to ", TextColors.GOLD, symbol, amount));
+        sender.sendMessage(Text.of(TextColors.GRAY, "You set ", recipient.getName(), "\'s balance to ", TextColors.GOLD, defaultCurrency.format(amount)));
 
         return CommandResult.success();
     }