+++ src/main/java/com/erigitic/config/TEAccount.java
@@ -14,7 +14,7 @@ import java.math.BigDecimal;
 import java.util.*;
 
 /**
- * Created by Erigitic on 1/1/2016.
+ * Created by Eric on 1/1/2016.
  */
 public class TEAccount implements UniqueAccount {
 
@@ -119,6 +119,14 @@ public class TEAccount implements UniqueAccount {
             BigDecimal newBalance = curBalance.add(amount);
 
             accountConfig.getNode(uuid.toString(), currencyName + "-balance").setValue(newBalance.setScale(2, BigDecimal.ROUND_DOWN));
+
+            // Reset balance to the money cap if it goes over
+            if (totalEconomy.isLoadMoneyCap()) {
+                if (getBalance(currency, contexts).compareTo(totalEconomy.getMoneyCap()) == 1) {
+                    accountConfig.getNode(uuid.toString(), currencyName + "-balance").setValue(totalEconomy.getMoneyCap());
+                }
+            }
+            
             accountManager.saveAccountConfig();
 
             transactionResult = new TETransactionResult(this, currency, amount, contexts, ResultType.SUCCESS, TransactionTypes.DEPOSIT);