+++ src/main/java/com/erigitic/config/TECurrency.java
@@ -4,9 +4,11 @@ import org.spongepowered.api.service.economy.Currency;
 import org.spongepowered.api.text.Text;
 
 import java.math.BigDecimal;
+import java.text.NumberFormat;
+import java.util.Locale;
 
 /**
- * Created by Erigitic on 1/1/2016.
+ * Created by Eric on 1/1/2016.
  */
 public class TECurrency implements Currency {
 
@@ -51,7 +53,7 @@ public class TECurrency implements Currency {
 
     @Override
     public Text format(BigDecimal amount, int numFractionDigits) {
-        return Text.of(amount.setScale(numFractionDigits, BigDecimal.ROUND_HALF_UP));
+        return Text.of(symbol, NumberFormat.getInstance(Locale.ENGLISH).format(amount.setScale(numFractionDigits, BigDecimal.ROUND_HALF_UP)));
     }
 
     @Override